+++ src/main/java/es/in2/dome/blockchain/connector/integration/contextbroker/configuration/ContextBrokerConfigApi.java
@@ -16,4 +16,5 @@ import org.springframework.context.annotation.Configuration;
 public class ContextBrokerConfigApi {
     private String subscriptionUrl;
     private String entitiesUrl;
+    private String accessToken;
 }